% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\docType{data}
\name{cases}
\alias{cases}
\title{Data on cases}
\format{
A data frame with 47 variables:
\describe{
\item{key_id}{Numeric. An ID number that uniquely identifies each
observation in the dataset. }
\item{case_id}{String. An ID number that uniquely identifies each
infringement case.}
\item{case_number}{Numeric. The number of the case.}
\item{case_year}{Numeric. The year of the case.}
\item{member_state_id}{Numeric. An ID number that uniquely identifies each
member state. This ID number is assigned when member states are sorted by
accession date and then alphabetically. }
\item{member_state}{String. The name of the member state that the
Commission opened the case against. }
\item{member_state_code}{String. A two letter code assigned by the
Commission that uniquely identifies each member state. }
\item{department_id}{Numeric. An ID number that uniquely identifies each
Commission department.}
\item{department}{String. The name of the Commission department that opened
the infringement case.}
\item{department_code}{String. A multi-letter code assigned by the
Commission that uniquely identifies each department.}
\item{directive}{Dummy. A dummy variable indicating whether the description
of the infringement case provided by the Commission references a directive
that the case relates to. Directive numbers have the format
\code{####/####}, where the first set of numbers is the year the directive
was enacted and the second set is the number of the directive.}
\item{directive_number}{String. The number of the directive that the
infringement case relates to, if applicable. Coded \code{NA} if the
description of the case provided by the Commission does not reference a
specific directive number.}
\item{celex}{String. The CELEX number of the directive that the
infringement case relates to, if applicable. Coded \code{NA} if the
description of the case provided by the Commission does not reference a
specific directive number. CELEX numbers for directives have the format
\code{3####L####}, where the first set of numbers is the year the directive
was enacted and the second set is the number of the directive.}
\item{case_type_id}{Numeric. An ID number that uniquely identifies each
type of state aid cases. Coded \code{1} for noncommunication cases, which
are cases that relate to a member state failing to notify the Commission
that it has transposed a directive by the stated deadline. Coded \code{2}
for nonconformity cases, which are cases that relate to a member state
incorrectly transposing a directive. }
\item{case_type}{String. The type of the infringement case. There are two
types of cases. Coded \code{noncommunication} for cases that relate to a
member state failing to notify the Commission that it has transposed a
directive by the stated deadline. Coded \code{nonconformity} for cases that
relate to a member state incorrectly transposing a directive. }
\item{noncommunication}{Dummy. A dummy variable indicating whether the
infringement case is a noncommunication case, which is a case that relates
to a member state failing to notify that Commission that it has transposed
a directive by the stated deadline.}
\item{nonconformity}{Dummy. A dummy variable indicating whether the
infringement case is a nonconformity case, which is a case that relates to
a member state incorrectly transposing a directive.}
\item{complete}{Dummy. A dummy variable indicating whether the public
records published by the Commission about the case are complete in that
they include a record of a letter of formal notice under Article 258 of the
Treaty of the Functioning of the European Union (TFEU), a record of the
closing of the case, and a record of any stages of the infringement
procedure in between. }
\item{count_press_releases}{Numeric. The number of press releases published
by the Commission related to the case.}
\item{stage_lfn_258}{Dummy. A dummy variable indicating whether the
decision is a letter of formal notice sent by the Commission under Article
258 of the Treaty on the Functioning of the European Union (TFEU).}
\item{stage_ro_258}{Dummy. A dummy variable indicating whether the decision
is a reasoned opinion sent by the Commission under Article 258 of the
Treaty on the Functioning of the European Union (TFEU).}
\item{stage_rf_258}{Dummy. A dummy variable indicating whether the decision
is a referral to the Court of Justice of the European Union (CJEU) made by
the Commission under Article 258 of the Treaty on the Functioning of the
European Union (TFEU).}
\item{stage_lfn_260}{Dummy. A dummy variable indicating whether the
decision is a letter of formal notice sent by the Commission under Article
260 of the Treaty on the Functioning of the European Union (TFEU).}
\item{stage_ro_260}{Dummy. A dummy variable indicating whether the decision
is a reasoned opinion sent by the Commission under Article 260 of the
Treaty on the Functioning of the European Union (TFEU).}
\item{stage_rf_260}{Dummy. A dummy variable indicating whether the decision
is a referral to the Court of Justice of the European Union (CJEU) made by
the Commission under Article 260 of the Treaty on the Functioning of the
European Union (TFEU).}
\item{stage_closing}{Dummy. A dummy variable indicating whether the
decision is the closing of the case by the Commission. }
\item{stage_withdrawal}{Dummy. A dummy variable indicating whether the
decision is the withdrawal of the case by the Commission.}
\item{case_history}{String. A summary of the history of the case, including
all decisions taken by the Commission in the case (with dates). }
\item{corrected_lfn_258}{Dummy. A dummy variable indicating whether the
decision is a letter of formal notice sent by the Commission under Article
258 of the Treaty on the Functioning of the European Union (TFEU).}
\item{corrected_ro_258}{Dummy. A dummy variable indicating whether the
decision is a reasoned opinion sent by the Commission under Article 258 of
the Treaty on the Functioning of the European Union (TFEU).}
\item{corrected_rf_258}{Dummy. A dummy variable indicating whether the
decision is a referral to the Court of Justice of the European Union (CJEU)
made by the Commission under Article 258 of the Treaty on the Functioning
of the European Union (TFEU).}
\item{corrected_lfn_260}{Dummy. A dummy variable indicating whether the
decision is a letter of formal notice sent by the Commission under Article
260 of the Treaty on the Functioning of the European Union (TFEU).}
\item{corrected_ro_260}{Dummy. A dummy variable indicating whether the
decision is a reasoned opinion sent by the Commission under Article 260 of
the Treaty on the Functioning of the European Union (TFEU).}
\item{corrected_rf_260}{Dummy. A dummy variable indicating whether the
decision is a referral to the Court of Justice of the European Union (CJEU)
made by the Commission under Article 260 of the Treaty on the Functioning
of the European Union (TFEU). Corrected based on data about decisions made
by the Commission at later stages of the infringement process.}
\item{count_decisions}{Numeric. The number of decisions issued by the
Commission in the case. Based on the corrected data.}
\item{count_lfn_258}{Numeric. The number of letters of formal notice sent
by the Commission under Article 258 of the Treaty on the Functioning of the
European Union (TFEU).}
\item{count_ro_258}{Numeric. The number of reasoned opinions sent by the
Commission under Article 258 of the Treaty on the Functioning of the
European Union (TFEU).}
\item{count_rf_258}{Numeric. The number of referrals to the Court of
Justice of the European Union (CJEU) made by the Commission under Article
258 of the Treaty on the Functioning of the European Union (TFEU).}
\item{count_lfn_260}{Numeric. The number of letters of formal notice sent
by the Commission under Article 260 of the Treaty on the Functioning of the
European Union (TFEU).}
\item{count_ro_260}{Numeric. The number of reasoned opinions sent by the
Commission under Article 260 of the Treaty on the Functioning of the
European Union (TFEU).}
\item{count_rf_260}{Numeric. The number of referrals to the Court of
Justice of the European Union (CJEU) made by the Commission under Article
260 of the Treaty on the Functioning of the European Union (TFEU).}
\item{count_missing}{Numeric. The number of decision stages prior to the
last observed decision stage that are missing.}
\item{missing_lfn_258}{Dummy. A dummy variable indicating whether the
Commission does not report a letter of formal notice under Article 258 of
the Treaty on the Functioning of the European Union (TFEU) but does report
a decision at a later stage of the infringement process.}
\item{missing_ro_258}{Dummy. A dummy variable indicating whether the
Commission does not report a reasoned opinion under Article 258 of the
Treaty on the Functioning of the European Union (TFEU) but does report a
decision at a later stage of the infringement process.}
\item{missing_rf_258}{Dummy. A dummy variable indicating whether the
Commission does not report a referral to the Court of Justice of the
European Union (CJEU) under Article 258 of the Treaty on the Functioning of
the European Union (TFEU) but does report a decision at a later stage of
the infringement process.}
\item{missing_lfn_260}{Dummy. A dummy variable indicating whether the
Commission does not report a letter of formal notice under Article 260 of
the Treaty on the Functioning of the European Union (TFEU) but does report
a decision at a later stage of the infringement process.}
\item{missing_ro_260}{Dummy. A dummy variable indicating whether the
Commission does not report a reasoned opinion under Article 260 of the
Treaty on the Functioning of the European Union (TFEU) but does report a
decision at a later stage of the infringement process.}
}
}
\usage{
cases
}
\description{
This dataset include data on infringement cases. There is one observation
per case (2002-2020). The dataset includes information on the department
responsible for the case, the member state that the case is against, and
the progression of the case through the infringement procedure.
}
\keyword{datasets}
