% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases_csts_ms_ct.R
\docType{data}
\name{cases_csts_ms_ct}
\alias{cases_csts_ms_ct}
\title{Case-level cross-sectional time-series data by member state and case type}
\format{
A data frame with 8 variables:
\describe{
\item{key_id}{\code{numeric}. An ID number that uniquely identifies each
observation in the dataset. }
\item{year}{\code{numeric}. The year the case was opened by the
Commission.}
\item{member_state_id}{\code{numeric}. An ID number that uniquely
identifies each member state. This ID number is assigned when member states
are sorted by accession date and then alphabetically. }
\item{member_state}{\code{string}. The name of the member state that the
Commission opened the case against. }
\item{member_state_code}{\code{string}. A two letter code assigned by the
Commission that uniquely identifies each member state. }
\item{case_type_id}{\code{numeric}. An ID number that uniquely identifies
each type of state aid cases. Coded \code{1} for noncommunication cases,
which are cases that relate to a member state failing to notify the
Commission that it has transposed a directive by the stated deadline. Coded
\code{2} for nonconformity cases, which are cases that relate to a member
state incorrectly transposing a directive. }
\item{case_type}{\code{string}. The type of the infringement case. There
are two types of cases. Coded \code{noncommunication} for cases that relate
to a member state failing to notify the Commission that it has transposed a
directive by the stated deadline. Coded \code{nonconformity} for cases that
relate to a member state incorrectly transposing a directive. }
\item{count_cases}{\code{numeric}. A count of the number of cases opened by
the Commission at this level of aggregation.}
}
}
\usage{
cases_csts_ms_ct
}
\description{
This dataset includes aggregated data on the number of infringement cases
per member state per year (cross-sectional time-series data) broken down by
case type (noncommunication vs nonconformity). There is one observation per
member state per year per case type (2002-2020), excluding state-years
where the state was not a member of the EU.
}
\keyword{datasets}
