% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases_ddy_d.R
\docType{data}
\name{cases_ddy_d}
\alias{cases_ddy_d}
\title{DDY data on cases (by case type)}
\format{
A data frame with 11 variables:
\describe{
\item{key_id}{Numeric. An ID number that uniquely identifies each
observation in the dataset. }
\item{year}{Numeric. The year the case was opened by the Commission.}
\item{department_id}{Numeric. An ID number that uniquely identifies each
Commission department.}
\item{department}{String. The name of the Commission department that opened
the infringement case.}
\item{department_code}{String. A multi-letter code assigned by the
Commission that uniquely identifies each department.}
\item{member_state_id}{Numeric. An ID number that uniquely identifies each
member state. This ID number is assigned when member states are sorted by
accession date and then alphabetically. }
\item{member_state}{String. The name of the member state that the
Commission opened the case against. }
\item{member_state_code}{String. A two letter code assigned by the
Commission that uniquely identifies each member state. }
\item{case_type_id}{Numeric. An ID number that uniquely identifies each
type of state aid cases. Coded \code{1} for noncommunication cases, which
are cases that relate to a member state failing to notify the Commission
that it has transposed a directive by the stated deadline. Coded \code{2}
for nonconformity cases, which are cases that relate to a member state
incorrectly transposing a directive. }
\item{case_type}{String. The type of the infringement case. There are two
types of cases. Coded \code{noncommunication} for cases that relate to a
member state failing to notify the Commission that it has transposed a
directive by the stated deadline. Coded \code{nonconformity} for cases that
relate to a member state incorrectly transposing a directive. }
\item{count_cases}{Numeric. A count of the number of cases opened by the
Commission at this level of aggregation.}
}
}
\usage{
cases_ddy_d
}
\description{
This dataset includes aggregated data on the number of infringement cases
per department per member state per year (directed dyad-year data) broken
down by case type (noncommunication vs nonconformity). There is one
observation per department per member state per year per case type
(2002-2020), excluding directed dyad-years where the state was not a member
of the EU. The dataset uses current department names.
}
\keyword{datasets}
