% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases_ts_d.R
\docType{data}
\name{cases_ts_d}
\alias{cases_ts_d}
\title{TS data on cases (by case type)}
\format{
A data frame with 5 variables:
\describe{
\item{key_id}{Numeric. An ID number that uniquely identifies each
observation in the dataset. }
\item{year}{Numeric. The year the case was opened by the Commission.}
\item{case_type_id}{Numeric. An ID number that uniquely identifies each
type of state aid cases. Coded \code{1} for noncommunication cases, which
are cases that relate to a member state failing to notify the Commission
that it has transposed a directive by the stated deadline. Coded \code{2}
for nonconformity cases, which are cases that relate to a member state
incorrectly transposing a directive. }
\item{case_type}{String. The type of the infringement case. There are two
types of cases. Coded \code{noncommunication} for cases that relate to a
member state failing to notify the Commission that it has transposed a
directive by the stated deadline. Coded \code{nonconformity} for cases that
relate to a member state incorrectly transposing a directive. }
\item{count_cases}{Numeric. A count of the number of cases opened by the
Commission at this level of aggregation.}
}
}
\usage{
cases_ts_d
}
\description{
This dataset includes aggregated data on the number of infringement cases
per year (time-series data) broken down by case type (noncommunication vs
nonconformity). There is one observation per year per case type
(2002-2020).
}
\keyword{datasets}
