% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisions_ddy.R
\docType{data}
\name{decisions_ddy}
\alias{decisions_ddy}
\title{DDY data on decisions}
\format{
A data frame with 11 variables:
\describe{
\item{key_id}{Numeric. An ID number that uniquely identifies each
observation in the dataset. }
\item{year}{Numeric. The year the decision was issued by the Commission.}
\item{department_id}{Numeric. An ID number that uniquely identifies each
Commission department.}
\item{department}{String. The name of the Commission department that opened
the infringement case.}
\item{department_code}{String. A multi-letter code assigned by the
Commission that uniquely identifies each department.}
\item{member_state_id}{Numeric. An ID number that uniquely identifies each
member state. This ID number is assigned when member states are sorted by
accession date and then alphabetically. }
\item{member_state}{String. The name of the member state that the
Commission opened the case against. }
\item{member_state_code}{String. A two letter code assigned by the
Commission that uniquely identifies each member state. }
\item{decision_stage_id}{Numeric. An ID number that uniquely identifies
each decision stage in the infringement procedure. Coded \code{1} for
letters of formal notice under Article 258 of the Treaty on the Functioning
of the European Union (TFEU), coded \code{2} for reasoned opinions under
Article 258, coded \code{3} for referrals to the Court under Article 258,
coded \code{4} for letters of formal notice under Article 260, coded
\code{5} for reasoned opinions under Article 260, and coded \code{6} for
referrals to the Court under Article 267}
\item{decision_stage}{String. The decision stage of the infringement
procedure. Possible values include: \code{Letter of formal notice (Article
258)}, \code{Reasoned opinion (Article 259)}, \code{Referral to the Court
(Article 258)}, \code{Letter of formal notice (Article 260)},
\code{Reasoned opinion (Article 260)}, and \code{Referral to the Court
(Article 260)}. }
\item{count_decisions}{Numeric. A count of the number of decisions made by
the Commission in infringement cases at this level of aggregation.}
}
}
\usage{
decisions_ddy
}
\description{
This dataset includes aggregated data on the number of decisions per stage
of the infringement procedure per department per member state per year
(directed dyad-year data). There is one observation per department per
member state per year per decision stage (2002-2020), excluding directed
dyad-years where the state was not a member of the EU. The dataset uses
current department names.
}
\keyword{datasets}
