% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisions_ts_d.R
\docType{data}
\name{decisions_ts_d}
\alias{decisions_ts_d}
\title{TS data on decisions (by case type)}
\format{
A data frame with 7 variables:
\describe{
\item{key_id}{Numeric. An ID number that uniquely identifies each
observation in the dataset. }
\item{year}{Numeric. The year the decision was issued by the Commission.}
\item{case_type_id}{Numeric. An ID number that uniquely identifies each
type of state aid cases. Coded \code{1} for noncommunication cases, which
are cases that relate to a member state failing to notify the Commission
that it has transposed a directive by the stated deadline. Coded \code{2}
for nonconformity cases, which are cases that relate to a member state
incorrectly transposing a directive. }
\item{case_type}{String. The type of the infringement case. There are two
types of cases. Coded \code{noncommunication} for cases that relate to a
member state failing to notify the Commission that it has transposed a
directive by the stated deadline. Coded \code{nonconformity} for cases that
relate to a member state incorrectly transposing a directive. }
\item{decision_stage_id}{Numeric. An ID number that uniquely identifies
each decision stage in the infringement procedure. Coded \code{1} for
letters of formal notice under Article 258 of the Treaty on the Functioning
of the European Union (TFEU), coded \code{2} for reasoned opinions under
Article 258, coded \code{3} for referrals to the Court under Article 258,
coded \code{4} for letters of formal notice under Article 260, coded
\code{5} for reasoned opinions under Article 260, and coded \code{6} for
referrals to the Court under Article 262}
\item{decision_stage}{String. The decision stage of the infringement
procedure. Possible values include: \code{Letter of formal notice (Article
258)}, \code{Reasoned opinion (Article 259)}, \code{Referral to the Court
(Article 258)}, \code{Letter of formal notice (Article 260)},
\code{Reasoned opinion (Article 260)}, and \code{Referral to the Court
(Article 260)}. }
\item{count_cases}{Numeric. A count of the number of cases opened by the
Commission at this level of aggregation.}
}
}
\usage{
decisions_ts_d
}
\description{
This dataset includes aggregated data on the number of decisions per stage
of the infringement procedure per year (time-series data) broken down by
case type (nonommunication vs nonconformity). There is one observation per
year per decision stage per case type (2002-2020).
}
\keyword{datasets}
