% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lists.R
\name{dg_list}
\alias{dg_list}
\title{Data on Directorates-General}
\usage{
dg_list()
}
\value{
Loads a tibble called \code{DGs} with the following variables into the global environment:
\itemize{
\item \code{directorate_general}: the name of the DG.
\item \code{directorate_general_code}: a short character code for the DG.
\item \code{type}: either \code{"Directorate-General"} or \code{"Service Department"}.
\item \code{status}: either \code{"current"} or \code{"former"}.
}
}
\description{
A function that laods a tibble with a list of Directorates-General (DGs).
}
