% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{EUIP_decisions_API}
\alias{EUIP_decisions_API}
\title{Query the Cases Dataset of the EUIP Database}
\usage{
EUIP_decisions_API(parameters)
}
\arguments{
\item{parameters}{A named list. The names of the elements of the list should
correspond exactly to parameters of the EUIP Database API. With the
exception of the \code{limit} and \code{offset} parameters, each parameter
corresponds to a variable from the dataset. See the EUIP Database API
Codebook for a list of valid parameters and the corresponding variable in
the dataset, if applicable. See the EUIP Database Codebook for a
description of the variables. All API parameters are in \code{lowerCamelCase}
with no exceptions. With the exception of the \code{limit} and \code{offset}
parameters, value of the elements of the list are the values that will be
used to filter the data. If specified, the \code{limit} parameter indicates the
maximum number of observations to return, and the \code{offset} parameter
indicates which observation of the filtered data to start at. If you
include invalid parameters, the function will still run, but you will get a
warning indicating which parameters are invalid.}
}
\description{
A function to query the \code{EUIP_decisions} dataset of the EUIP Database using
the EUIP Database API.
}
