% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_decision_data}
\alias{get_decision_data}
\title{Get Data on EU Infringement Decisions}
\usage{
get_decision_data(
  id = NULL,
  year = NULL,
  date = NULL,
  ms = NULL,
  dg = NULL,
  dir = NULL,
  celex = NULL,
  type = NULL,
  stage = NULL
)
}
\arguments{
\item{id}{A numeric vector where each number is a valid case number.
Case numbers always have 8 digits.
The first 4 digits are the year the case was opened and the second 4 digits uniquely identify the cases within the year.
If you include one or more invalid case numbers, you will get an error.}

\item{year}{A four-digit numeric value indicating a year or a numeric vector of length 2 containing a start year and an end year.
If you list more than 2 years, or do not list the start year and end year in chronological order, you will get an error.
The EUIP dataset covers the period from 2002 to 2020.
If you include a year outside this range, you will also get an error.}

\item{ms}{A string vector where each string is a valid member state name or code.
Use the function \code{ms_list()} to get a complete list of valid member state names and codes.
The API will return an error if one or more of the elements is not valid.}

\item{dg}{A string vector where each string is a valid DG name or code.
Use the function \code{dg_list()} to get a complete list of valid DG names and codes.
The API will return an error if one or more of the elements is not valid.}

\item{dir}{A vector of valid directive numbers.
Note that not all observations correspond to a specific directive.
If a observation does not, and this parameter is not \code{NULL}, the API will not return it,
even if it meets all of the other specified criteria.
All observations related to noncommunication cases should correspond to a specific directive,
but the Commission does not always report this information.
If an invalid directive number is included, the API will return an error.}

\item{celex}{A vector of valid CELEX numbers for directives.
Note that not all observations correspond to a specific directive.
If a observation does not, and this parameter is not \code{NULL}, the API will not return it,
even if it meets all of the other specified criteria.
All observations related to noncommunication cases should correspond to a specific directive,
but the Commission does not always report this information.
If an invalid CELEX number is included, the API will return an error.}

\item{type}{A string vector where each string is a type of infringment case.
Valid options include:
\itemize{
\item \code{"COMM"}, for noncommunication cases; and
\item \code{"CONF"}, for nonconformity cases.
}
If \code{NULL}, the function will return decisions in both types of cases.
Noncommunication cases are those in which the Commission alleges that a member state
has failed to notify the Commission of the transposition of a directive before the specified date.
Nonconformity cases are those in which the Commission alleges that a member state
has failed to correctly transpose a directive into national law.}

\item{stage}{A string vector where each string is a stage in the infringement procedure.
Only decisions in cases that make it to all specified stages will be included in the results.
Valid options include:
\itemize{
\item \code{"LFN258"}, for letters of formal notice under Article 258 (TFEU);
\item \code{"RO258"}, for reasoned opinions under Article 258 (TFEU);
\item \code{"RF258"}, for referrals to the Court under Article 258 (TFEU);
\item \code{"LFN260"}, for letters of formal notice under Article 260 (TFEU);
\item \code{"RO260"}, for reasoned opinions under Article 260 (TFEU);
\item \code{"RF260"}, for referrals to the Court under Article 260 (TFEU);
\item \code{"C"}, for the closing of the case; and
\item \code{"W"}, for the withdrawal of the case.
}
Note that a case can be closed or withdrawn, but not both.}
}
\value{
Returns a tibble in tidy format with decision-level data, subject to the specified filters.
The unit of observation is a decision by the Commission in an infringement case.
The tibble will include the following variables:
\itemize{
\item \code{case_number}: a unique ID number assigned to the case by the Commission;
\item \code{decision_date}: the date the decision was made in \code{YYYY-MM-DD} format;
\item \code{decision_date}: the year the decision was madet;
\item \code{member_state}: the name of the member state under investigation;
\item \code{member_state_code}: the code of the member state under investigation;
\item \code{directorate_general}: the name of the DG that managed the case;
\item \code{directorate_general_code}: the code of the DG that managed the case;
\item \code{case_type}: either \code{"Noncommunication"} or \code{"Nonconformity"}; and
\item \code{decision_stage}: the type of the decision (letter of formal notice, reasoned opinion, etc.).
}
}
\description{
This function provides an easy way to get decision-level data from the European Union Infringement Procedures (EUIP)
dataset API without writing an API query and parsing the JSON data returned by the API.
This function returns one observation per decision taken by the European Commission in
an infringement case that matches all of the search criteria you specify.
}
\details{
The EUIP dataset API returns data in a standard JSON format.
This function automatically converts the JSON data into a tibble (see the \code{tidyverse} package).
The output of this function will be in a tidy format
(each variable in its own column, each observation in its own row, and each value in its own cell)
for easy use with other functions in the \code{tidyverse}, such as \code{ggplot}.
}
\seealso{
Other functions in the \code{euinfr} package: \code{\link[=get_case_data]{get_case_data()}} for case-level data,
\code{\link[=get_ms_data]{get_ms_data()}} for time-varying member state-level data, and
\code{\link[=get_dg_data]{get_dg_data()}} for time-varying DG-level data.
}
