% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_dg_data}
\alias{get_dg_data}
\title{Get Directorate-General Data}
\usage{
get_dg_data(year = NULL, dg = NULL, type = NULL, stage = NULL)
}
\arguments{
\item{year}{A four-digit numeric value indicating a year or a numeric vector of length 2 containing a start year and an end year.
If you list more than 2 years, or do not list the start year and end year in chronological order, you will get an error.
The EUIP dataset covers the period from 2002 to 2020.
If you include a year outside this range, you will also get an error.}

\item{dg}{A string vector where each string is a valid DG name or code.
Use the function \code{dg_list()} to get a complete list of valid DG names and codes.
The API will return an error if one or more of the elements is not valid.}

\item{type}{A string vector where each string is a type of infringment case.
Valid options include:
\itemize{
\item \code{"COMM"}, for noncommunication cases; and
\item \code{"CONF"}, for nonconformity cases.
}
If \code{NULL}, the function will not distinguish between noncommunication cases and nonconformity cases.
Noncommunication cases are those in which the Commission alleges that a member state
has failed to notify the Commission of the transposition of a directive before the specified date.
Nonconformity cases are those in which the Commission alleges that a member state
has failed to correctly transpose a directive into national law.}

\item{stage}{A string vector where each string is a stage in the infringement procedure.
Valid options include:
\itemize{
\item \code{"LFN258"}, for letters of formal notice under Article 258 (TFEU);
\item \code{"RO258"}, for reasoned opinions under Article 258 (TFEU);
\item \code{"RF258"}, for referrals to the Court under Article 258 (TFEU);
\item \code{"LFN260"}, for letters of formal notice under Article 260 (TFEU);
\item \code{"RO260"}, for reasoned opinions under Article 260 (TFEU); and
\item \code{"RF260"}, for referrals to the Court under Article 260 (TFEU).
}}
}
\value{
Returns a tibble in tidy format with time-varying DG-level data on the number decisions made by the Commission
in infringement cases at each stage of the infringement procedure (letter of formal notice, reasoned opinion, etc.)
and by type of case (e.g., noncommunication vs nonconformity), subject to the specified filters.
The unit of observation is the unique combination of a year, a DG, a type of case, and a stage of the infringement procedure.
The tibble will include the following variables:
\itemize{
\item \code{directorate_general}: the name of the member state;
\item \code{directorate_general_code}: a 2-character code for the member state;
\item \code{year}: the year of the decision (not the year the case was lodged);
\item \code{decision_stage}: the stage of the infringement procedure;
\item \code{case_type}: if \code{type} is not \code{NULL}, the type of case (\code{"Noncommunication"} or \code{"Nonconformity"}), otherwise excluded; and
\item \code{count}: the number of decisions in cases of the indicated type at the indicated stage
involving the indicated DG in the indicated year.
}
}
\description{
This function provides an easy way to get time-varying DG-level data from the European Union Infringement Procedures (EUIP)
dataset API without writing an API query and parsing the JSON data returned by the API.
This function returns time-varying DG-level data on the number decisions made by the Commission
in infringement cases at each stage of the infringement procedure (letter of formal notice, reasoned opinion, etc.)
with an option to disaggregate by type of case (e.g., noncommunication vs nonconformity).
Use the arguments to apply filters that narrow down the results to data for
specific years, specific DGs, specific types of cases, and/or specific stages of the infringement procedure.
}
\details{
The EUIP dataset API returns data in a standard JSON format.
This function automatically converts the JSON data into a tibble (see the \code{tidyverse} package).
The output of this function is in a tidy format
(each variable in its own column, each observation in its own row, and each value in its own cell)
for easy use with other functions in the \code{tidyverse}, such as \code{ggplot}.
}
\examples{
# the number of letters of formal notice under Article 258
# sent by DG ENV and DG ENER in noncommunication cases
# for each year from 2010 through 2010

get_dg_data(year = c(2010, 2020), dg = c("ENV", "ENER"), type = "COMM", stage = "LFN258")

# the number of decisions that the Commission made in cases
# managed by DG GROW by type in nonconformity cases in 2010

get_dg_data(year = 2019, dg = "GROW", type = "CONF")

}
\seealso{
Other functions in the \code{euinfr} package: \code{\link[=get_decision_data]{get_decision_data()}} for decision-level data, \code{\link[=get_case_data]{get_case_data()}} for case-level data, and
\code{\link[=get_ms_data]{get_ms_data()}} for time-varying member state-level data.
}
