% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lists.R
\name{ms_list}
\alias{ms_list}
\title{Data on Member States}
\usage{
ms_list()
}
\value{
Loads a tibble called \code{member_states} with the following variables into the global environment:
\itemize{
\item \code{member_state}: the name of the member state; and
\item \code{member_state_code}: a 2-character code for the member state.
}
}
\description{
A function that loads a tibble with a list of EU member states into the global environment.
}
